Add files via upload
The file island2.py will work well, but the line 102 is called even though its not required. But the answers are correct.
Also it is to be noted that the first matched is gen to max load, the reverse is being implemented in the first step. There while sort, sort=reverse has to be done. Then I guess it will work perfect.