Add files via upload
Kindly check the loop is not implementing properly. What is the property of the a.pop(0). Here the value is not assigned to a node even though it should have been. The islands_all is updating. But the values in the bus[][3] should be updating, but its not getting updated.